+++ Users.js
@@ -56,7 +56,6 @@ class Users extends React.Component {
       connect: PropTypes.func.isRequired,
       hasPerm: PropTypes.func.isRequired,
     }).isRequired,
-    data: PropTypes.object.isRequired,
     resources: PropTypes.shape({
       users: PropTypes.shape({
         hasLoaded: PropTypes.bool.isRequired,
@@ -191,7 +190,8 @@ class Users extends React.Component {
   }
 
   componentWillUpdate() {
-    const pg = this.props.data.patronGroups;
+    const pg = (this.props.resources.patronGroups || {}).records || [];
+
     if (pg && pg.length) {
       filterConfig[1].values = pg.map(rec => ({ name: rec.group, cql: rec.id }));
     }
@@ -259,7 +259,7 @@ class Users extends React.Component {
   }
 
   onNeedMore = () => {
-    this.props.mutator.userCount.replace(this.props.data.userCount + RESULT_COUNT_INCREMENT);
+    this.props.mutator.userCount.replace(this.props.resources.userCount + RESULT_COUNT_INCREMENT);
   }
 
   getRowURL(rowData) {
@@ -275,18 +275,18 @@ class Users extends React.Component {
     this.transitionToParams({ filters: Object.keys(filters).filter(key => filters[key]).join(',') });
   }
 
-  create = (data) => {
-    if (data.personal.addresses) {
-      data.personal.addresses = toUserAddresses(data.personal.addresses, this.props.data.addressTypes); // eslint-disable-line no-param-reassign
+  create = (user) => {
+    if (user.personal.addresses) {
+      user.personal.addresses = toUserAddresses(user.personal.addresses, this.props.user.addressTypes); // eslint-disable-line no-param-reassign
     }
 
     // extract creds object from user object
-    const creds = Object.assign({}, data.creds, { username: data.username });
-    if (data.creds) delete data.creds; // eslint-disable-line no-param-reassign
+    const creds = Object.assign({}, user.creds, { username: user.username });
+    if (user.creds) delete user.creds; // eslint-disable-line no-param-reassign
     // POST user record
-    this.props.mutator.users.POST(data);
+    this.props.mutator.users.POST(user);
     // POST credentials, permission-user, permissions;
-    this.postCreds(data.username, creds);
+    this.postCreds(user.username, creds);
     this.onClickCloseNewUser();
   }
 
@@ -352,8 +352,10 @@ class Users extends React.Component {
   }
 
   render() {
-    const { data, resources, stripes } = this.props;
+    const { resources, stripes } = this.props;
     const users = (resources.users || {}).records || [];
+    const patronGroups = (resources.patronGroups || {}).records || [];
+    const addressTypes = (resources.addressTypes || {}).records || [];
 
     /* searchHeader is a 'custom pane header'*/
     const searchHeader = <FilterPaneSearch searchFieldId="input-user-search" onChange={this.onChangeSearch} onClear={this.onClearSearch} resultsList={this.resultsList} value={this.state.searchTerm} placeholder={'Search'} />;
@@ -375,7 +377,7 @@ class Users extends React.Component {
       Name: user => `${_.get(user, ['personal', 'lastName'], '')}, ${_.get(user, ['personal', 'firstName'], '')}`,
       Barcode: user => user.barcode,
       'Patron Group': (user) => {
-        const pg = this.props.data.patronGroups.filter(g => g.id === user.patronGroup)[0];
+        const pg = patronGroups.filter(g => g.id === user.patronGroup)[0];
         return pg ? pg.group : '?';
       },
       Username: user => user.username,
@@ -386,7 +388,7 @@ class Users extends React.Component {
       this.props.stripes.hasPerm('users.item.get') ?
         (<Route
           path={`${this.props.match.path}/view/:userid/:username`}
-          render={props => <this.connectedViewUser stripes={stripes} okapi={this.okapi} paneWidth="44%" onClose={this.collapseDetails} addressTypes={data.addressTypes} {...props} />}
+          render={props => <this.connectedViewUser stripes={stripes} okapi={this.okapi} paneWidth="44%" onClose={this.collapseDetails} addressTypes={addressTypes} {...props} />}
         />) :
         (<div
           style={{
@@ -451,15 +453,15 @@ class Users extends React.Component {
         </Pane>
 
         {detailsPane}
-        <Layer isOpen={data.addUserMode ? data.addUserMode.mode : false} label="Add New User Dialog">
+        <Layer isOpen={resources.addUserMode ? resources.addUserMode.mode : false} label="Add New User Dialog">
           <UserForm
             id="userform-adduser"
             initialValues={{ active: true, personal: { preferredContactTypeId: '002' } }}
-            addressTypes={data.addressTypes}
+            addressTypes={addressTypes}
             onSubmit={(record) => { this.create(record); }}
             onCancel={this.onClickCloseNewUser}
             okapi={this.okapi}
-            optionLists={{ patronGroups: this.props.data.patronGroups, contactTypes }}
+            optionLists={{ patronGroups, contactTypes }}
           />
         </Layer>
       </Paneset>