+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Made the diplay of the create new user button conditional (UIS-71).
 * Fetch up to 100 loan records rather then the default of 10. Fixes UIU-173.
 * Displayed number of loans is now stable (side-effect of UIU-173 fix). Fixes UIU-184.
+* Switch from props.data to props.resources. Fixes UIU-136.
 
 ## [2.9.0](https://github.com/folio-org/ui-users/tree/v2.9.0) (2017-08-03)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.8.0...v2.9.0)